+++ ICD.Connect.Protocol.Mock/Ports/IrPort/MockIrPort.cs
@@ -68,6 +68,7 @@ namespace ICD.Connect.Protocol.Mock.Ports.IrPort
 		/// <param name="command"></param>
 		protected override void PressFinal(string command)
 		{
+			PrintTx(command);
 		}
 
 		/// <summary>
@@ -75,6 +76,7 @@ namespace ICD.Connect.Protocol.Mock.Ports.IrPort
 		/// </summary>
 		protected override void ReleaseFinal()
 		{
+			PrintTx("Release");
 		}
 
 		#endregion