+++ ICD.Connect.Protocol.Network/Ports/Web/HttpPort.Standard.cs
@@ -20,6 +20,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		private readonly HttpClient m_Client;
 		private readonly HttpClientHandler m_ClientHandler;
 		private readonly SafeCriticalSection m_ClientBusySection;
+		private bool m_ClientBusy;
 
 		#region Properties
 
@@ -48,11 +49,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		{
 			get
 			{
-				if (!m_ClientBusySection.TryEnter())
-					return true;
-
-				m_ClientBusySection.Leave();
-				return false;
+				return m_ClientBusySection.Execute(() => m_ClientBusy);
 			}
 		}
 
@@ -106,7 +103,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 			if (headers == null)
 				throw new ArgumentNullException("headers");
 
-			m_ClientBusySection.Enter();
+			m_ClientBusySection.Execute(() => m_ClientBusy = true);
 
 			try
 			{
@@ -124,7 +121,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 			}
 			finally
 			{
-				m_ClientBusySection.Leave();
+				m_ClientBusySection.Execute(() => m_ClientBusy = false);
 			}
 		}
 
@@ -137,7 +134,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// <returns></returns>
 		public override WebPortResponse Post(string relativeOrAbsoluteUri, Dictionary<string, List<string>> headers, byte[] data)
 		{
-			m_ClientBusySection.Enter();
+			m_ClientBusySection.Execute(() => m_ClientBusy = true);
 
 			try
 			{
@@ -158,7 +155,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 			}
 			finally
 			{
-				m_ClientBusySection.Leave();
+				m_ClientBusySection.Execute(() => m_ClientBusy = false);
 			}
 		}
 
@@ -171,7 +168,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// <returns></returns>
 		public override WebPortResponse Patch(string relativeOrAbsoluteUri, Dictionary<string, List<string>> headers, byte[] data)
 		{
-			m_ClientBusySection.Enter();
+			m_ClientBusySection.Execute(() => m_ClientBusy = true);
 
 			try
 			{
@@ -192,7 +189,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 			}
 			finally
 			{
-				m_ClientBusySection.Leave();
+				m_ClientBusySection.Execute(() => m_ClientBusy = false);
 			}
 		}
 
@@ -205,7 +202,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// <returns></returns>
 		public override WebPortResponse Put(string relativeOrAbsoluteUri, Dictionary<string, List<string>> headers, byte[] data)
 		{
-			m_ClientBusySection.Enter();
+			m_ClientBusySection.Execute(() => m_ClientBusy = true);
 
 			try
 			{
@@ -226,7 +223,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 			}
 			finally
 			{
-				m_ClientBusySection.Leave();
+				m_ClientBusySection.Execute(() => m_ClientBusy = false);
 			}
 		}
 
@@ -242,7 +239,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 
 			Accept = SOAP_ACCEPT;
 
-			m_ClientBusySection.Enter();
+			m_ClientBusySection.Execute(() => m_ClientBusy = true);
 
 			try
 			{
@@ -261,7 +258,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 			}
 			finally
 			{
-				m_ClientBusySection.Leave();
+				m_ClientBusySection.Execute(() => m_ClientBusy = false);
 			}
 
 			SetLastRequestSucceeded(false);
@@ -281,7 +278,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		{
 			WebPortResponse output = WebPortResponse.Failed;
 
-			m_ClientBusySection.Enter();
+			m_ClientBusySection.Execute(() => m_ClientBusy = true);
 
 			try
 			{
@@ -337,7 +334,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 			}
 			finally
 			{
-				m_ClientBusySection.Leave();
+				m_ClientBusySection.Execute(() => m_ClientBusy = false);
 			}
 
 			SetLastRequestSucceeded(output.Success);